project24-aicompanion/ai-companion/whisper-service/requirements.txt
fastapi
uvicorn
openai-whisper
numpy
pydantic
python-multipart
ffmpeg-python
